bpost: Update to support new version of Bpost CSV files
Update to latest OCA coding standards
