+++ l10n_be_account_bank_statement_import_bpost/wizard/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import account_bank_statement_import
